split Template.php classes to new files
